Refactor data services and update README for B3 model
- Moved data loading and preprocessing services to a new directory structure under `src/b3/service/data/db/b3`.
- Introduced `B3DataLoader`, `B3DataLoadingService`, and `B3DataPreprocessingService` for improved data handling and preprocessing.
- Updated import paths in relevant files to reflect the new structure.
- Added new methods for data fetching and preprocessing in the respective services.
- Enhanced README.md with updated usage instructions for the new data services.
