+++ src/b3/service/web_api/b3_model_api.py
@@ -4,8 +4,8 @@ from dotenv import load_dotenv
 from flask import Flask
 from flask_cors import CORS
 
-from b3.service.data.data_loading_service import B3DataLoadingService
-from b3.service.data.data_preprocessing_service import B3DataPreprocessingService
+from b3.service.data.db.b3.data_loading_service import B3DataLoadingService
+from b3.service.data.db.b3.data_preprocessing_service import B3DataPreprocessingService
 from b3.service.model.model_evaluation_service import B3ModelEvaluationService
 from b3.service.model.model_saving_service import B3ModelSavingService
 from b3.service.model.model_training_service import B3ModelTrainingService
